feat: IconType 분리 (#19)
